requirements.txt
django
chromadb
langchain==0.3.7
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-groq
sentence-transformers
pypdf
pandas
python-dotenv
